Modernize FolioReaderKit: Remove Realm/Carthage, Refactor Center, and Integrate WebServer
- Removed Realm and Carthage dependencies and cleaned up project configs.
- Refactored FolioReaderCenter pagination logic into ReaderPaginationEngine.
- Unified JS-to-Swift communication with EpubJSBridge.
- Decoupled scroll and collection view delegation into ReaderScrollDelegateHandler.
- Integrated GCDWebServer into FolioReaderContainer to serve EPUB resources via HTTP.
- Updated Package.swift for iOS 13.0+ compatibility and corrected resource paths.
- Added comprehensive unit tests for models, providers, and parsing logic.
- Fixed multiple UI layout issues and compilation errors.

+++ Sources/FolioReaderKit/EPUBCore/FREpubParserArchive.swift
@@ -125,7 +125,7 @@ class FREpubParserArchive: NSObject {
             }
         }
 
-        var entrySizeMap = [String:Int]()
+        var entrySizeMap = [String:UInt64]()
         archive.forEach { entry in
             entrySizeMap[entry.path] = entry.uncompressedSize
         }
